move components from core to lite
