+++ @xen-orchestra/lite/src/components/pool/dashboard/AlarmRow.vue
@@ -57,7 +57,7 @@ import { parseDateTime } from '@/libs/utils'
 import type { RawObjectType } from '@/libs/xen-api/xen-api.types'
 import { rawTypeToType } from '@/libs/xen-api/xen-api.utils'
 import type { XenApiAlarm } from '@/types/xen-api'
-import ArrowCollapseButton from '@core/components/ArrowCollapseButton.vue'
+import ArrowCollapseButton from '@/components/pool/dashboard/ArrowCollapseButton.vue'
 import { useUiStore } from '@core/stores/ui.store'
 import { storeToRefs } from 'pinia'
 import { ref } from 'vue'
