+++ @xen-orchestra/lite/src/components/pool/dashboard/ArrowCollapseButton.vue
